fix: enable re-downloading of shared files
