+++ src/components/Shell/PeerDownloadFileButton.tsx
@@ -1,4 +1,3 @@
-import streamSaver from 'streamsaver'
 import Fab from '@mui/material/Fab'
 import Download from '@mui/icons-material/Download'
 
@@ -8,9 +7,6 @@ import { Peer } from 'models/chat'
 interface PeerDownloadFileButtonProps {
   peer: Peer
 }
-
-// FIXME: Make this configurable
-streamSaver.mitm = 'https://jeremyckahn.github.io/StreamSaver.js/mitm.html'
 
 export const PeerDownloadFileButton = ({
   peer,
@@ -22,20 +18,7 @@ export const PeerDownloadFileButton = ({
   }
 
   const handleDownloadFileClick = async () => {
-    const torrent = await torrentClient.add(torrentMetadata.magnetURI)
-
-    for (const file of torrent.files) {
-      const fileStream = streamSaver.createWriteStream(file.name)
-      const writer = fileStream.getWriter()
-      file
-        .createReadStream()
-        .on('data', data => {
-          writer.write(data)
-        })
-        .on('end', () => {
-          writer.close()
-        })
-    }
+    await torrentClient.download(torrentMetadata.magnetURI)
   }
 
   return (
